PostModal shows post and comments from ExplorePage
App refactored to handle route history properly since the PostModal should open with the /p/:postId but if closed it should go back to explore page. If refreshed page, the /p/:postId will show the post as a post page. This behavior required wrapping the app in <router> in order to handle history and conditionally set the previousLocation for the router based on PostModal state.